% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rivernetwork}
\alias{rivernetwork}
\title{Simulate a River Network}
\usage{
rivernetwork(
  n_points = 1000,
  xmn = 0,
  ymn = 0,
  scl = 1,
  progressbar = T,
  raster = F,
  nrow = 250,
  ncol = 250
)
}
\arguments{
\item{n_points}{integer indicating the number of randomly generated
points/nodes. The larger the number of points/nodes, the more complex the
resulting rivernetwork. Set to 1000 by default.}

\item{xmn}{numeric minimum x-coordiante (left border)}

\item{ymn}{numeric minimum y-coordiante (bottom border)}

\item{scl}{numeric scaling factor by which the final rivernetwork should be
scaled. By default, the network will have a square extent of size 1 x 1.}

\item{progressbar}{logical indicating if a progressbar should be printed. Set
to TRUE by default.}

\item{raster}{logical indicating if a raster should be returned (i.e. if the
river network should be rasterized)}

\item{nrow}{integer number of rows in case a raster should be returned. 250
by default.}

\item{ncol}{integer number of columns in case a raster should be returned.
250 by default.}
}
\value{
\code{SpatialPolygonsDataFrame} of the river network
}
\description{
This function is used to simulate a river network.
}
\examples{
# Generate rivernetwork
set.seed(123)
riv <- rivernetwork(n_points = 500)

# Visualize it
plot(riv, col = "cornflowerblue", border = NA)
axis(1)
axis(2)

# Repeat, but returning a raster instead of SPDF
riv <- rivernetwork(n_points = 500, raster = T)
plot(riv, col = c("white", "cornflowerblue"))
}
